pyproject.toml
[tool.poetry]
name = "credit_card_shift"
version = "0.1.0"
description = ""
authors = ["tulakin <a.tulakin@cft.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.3"
fastapi = "^0.100.0"
PyYAML = "^6.0"
pydantic-settings = "^2.0.2"
uvicorn = "^0.23.0"
email-validator = "^2.0.0"
contextvars = "^2.4"
msgpack = "^1.0.5"
pytz = "^2023.3"
prometheus-client = "^0.17.1"
dependency-injector = "^4.41.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.19"}
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dateutil = "^2.8.2"
aiohttp = "^3.8.5"

#migration deps
alembic = {version = "^1.11.1", optional = true}

# linter deps
wemake-python-styleguide = {version = "0.18.0", optional = true}
flake8-noqa = {version = "1.3.2", optional = true}

# test deps
pytest = {version = "^7.4.0", optional = true}
pytest-cov = {version = "^4.1.0", optional = true}
httpx = {version = "^0.24.1", optional = true}
pytest-asyncio = "^0.21.1"
asyncio = "^3.4.3"
factory-boy = "^3.3.0"
greenlet = "^3.0.0"

[tool.poetry.extras]
lint = ["flake8-noqa", "wemake-python-styleguide"]
tests = ["pytest", "pytest-cov", "httpx"]
migrations = ["alembic"]

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
